Blog level 1: 1. Jpa was added (dependency was injected) 2. The InvalidDefinitionException caused by findById method was solved by @JsonIgnoreProperties({"hibernateLazyInitializer"}) on entity class 3. Auto generation of the db was added

+++ src/main/java/com/luxcampus/Blog/service/PostServiceInterface.java
@@ -10,9 +10,9 @@ public interface PostServiceInterface {
 
     void save(Post post);
 
-    void delete(int id);
+    void delete(Long id);
 
-    void update(int id, Post post);
+    void update(Long id, Post post);
 
-    Post findById(int id);
+    Post findById(Long id);
 }
